优化项目结构

+++ bg-app/bg-oauth-center/bg-auth-server/src/main/java/com/quan/core/service/impl/SysClientServiceImpl.java
@@ -9,8 +9,12 @@ import com.quan.core.common.model.SysClient;
 import com.quan.core.common.web.JsonResult;
 import com.quan.core.common.web.PageResult;
 import com.quan.core.common.web.Result;
+import com.quan.core.constant.AuthErrorCode;
+import com.quan.core.controller.request.client.QueryClientPageRequest;
+import com.quan.core.controller.request.client.QueryClientRequest;
 import com.quan.core.dao.SysClientDao;
 import com.quan.core.dao.SysClientServiceDao;
+import com.quan.core.factory.AuthClientFactory;
 import com.quan.core.service.SysClientService;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.MapUtils;
@@ -22,74 +26,69 @@ import org.springframework.security.oauth2.provider.ClientDetails;
 import org.springframework.security.oauth2.provider.client.JdbcClientDetailsService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.support.TransactionTemplate;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 @Slf4j
 @Service
-@SuppressWarnings("all")
 public class SysClientServiceImpl implements SysClientService {
-
 
     @Autowired
     private SysClientDao sysClientDao;
-
     @Autowired
     private SysClientServiceDao sysClientServiceDao;
-
     @Autowired
     private RedisTemplate<String, Object> redisTemplate;
-
     @Autowired
     private PasswordEncoder passwordEncoder;
-
     @Autowired
     private JdbcClientDetailsService jdbcClientDetailsService;
-
-
-    @Override
-    public Result saveOrUpdate(SysClient sysClient) {
-        try {
-            sysClient.setClientSecret(passwordEncoder.encode(sysClient.getClientSecretStr()));
-
-            if (sysClient.getId() != null) {// 修改
-                sysClientDao.updateByPrimaryKey(sysClient);
-            } else {// 新增
-                SysClient r = sysClientDao.getClient(sysClient.getClientId());
-                if (r != null) {
-                    return JsonResult.failed(sysClient.getClientId() + "已存在");
-                }
-                sysClientDao.save(sysClient);
-            }
-            return JsonResult.succeed("操作成功");
-        } catch (Exception e) {
-            throw new ServiceException(e);
-        }
-    }
+    @Autowired
+    private TransactionTemplate template;
 
 
     @Override
     @Transactional
+    public Result saveOrUpdate(SysClient sysClient) {
+        sysClient.setClientSecret(passwordEncoder.encode(sysClient.getClientSecretStr()));
+        if (Objects.isNull(sysClient.getId())) {
+            SysClient client = sysClientDao.getClient(sysClient.getClientId());
+            if (Objects.nonNull(client)) {
+                return JsonResult.failed(AuthErrorCode.CLIENT_EXISTS_ERROR.getMsg());
+            }
+            sysClientDao.save(sysClient);
+        } else {
+            sysClientDao.updateByPrimaryKey(sysClient);
+        }
+        return JsonResult.succeed();
+    }
+
+
+    @Override
     public void delete(Long id) {
-        try {
+        template.execute(transactionStatus -> {
             SysClient client = sysClientDao.getById(id);
+            if (Objects.isNull(client)) {
+                throw new ServiceException(AuthErrorCode.CLIENT_NOT_EXISTS_ERROR.getMsg());
+            }
             sysClientDao.delete(id);
             sysClientServiceDao.delete(id, null);
             redisTemplate.boundHashOps(OAuthConstant.CACHE_CLIENT_KEY).delete(client.map().getClientId());
             log.debug("删除应用id:{}", id);
-        } catch (Exception e) {
-            throw new ServiceException(e);
-        }
+            return true;
+        });
     }
 
     @Override
-    public PageResult<SysClient> list(Map<String, Object> params) {
+    public PageResult<SysClient> list(QueryClientPageRequest client) {
 
         try {
             //设置分页信息，分别是当前页数和每页显示的总记录数【记住：必须在mapper接口中的方法执行之前设置该分页信息】
-            PageHelper.startPage(MapUtils.getInteger(params, "page"), MapUtils.getInteger(params, "limit"), true);
-            List<SysClient> list = sysClientDao.findList(params);
+            PageHelper.startPage(client.getPageNumber(), client.getPageSize(), true);
+            List<SysClient> list = sysClientDao.findList(AuthClientFactory.newInstance(client));
             PageInfo<SysClient> pageInfo = new PageInfo<>(list);
             return PageResult.<SysClient>builder().data(pageInfo.getList()).code(0).count(pageInfo.getTotal()).build();
         } catch (Exception e) {
@@ -107,8 +106,8 @@ public class SysClientServiceImpl implements SysClientService {
     }
 
     @Override
-    public List<SysClient> findList(Map<String, Object> params) {
-        return sysClientDao.findList(params);
+    public List<SysClient> data(QueryClientRequest client) {
+        return sysClientDao.data(AuthClientFactory.newInstance(client));
     }
